requirement.txt
asgiref==3.2.10
Django==3.1
feedparser==5.2.1
mysql-connector-python==8.0.21
mysqlclient==2.0.1
protobuf==3.13.0
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
